{"schema":"decodex/commit/1","summary":"Split Decodex app panel controls","authority":"XY-1115"}
