+++ apps/decodex-app/Sources/DecodexApp/PanelControls.swift
@@ -1,0 +1,145 @@
+import SwiftUI
+
+struct SummaryTileView: View {
+	let title: String
+	let value: String
+	let symbol: String
+	let tint: Color
+	@Environment(\.colorScheme) private var colorScheme
+
+	var body: some View {
+		HStack(alignment: .firstTextBaseline, spacing: 4) {
+			PanelMetricIconView(
+				symbol: symbol,
+				tint: tint.opacity(colorScheme == .dark ? 0.78 : 0.82)
+			)
+
+			Text(title)
+				.font(PanelFont.metricLabel)
+				.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.82))
+				.lineLimit(1)
+				.fixedSize(horizontal: true, vertical: false)
+
+			Text(value)
+				.font(PanelFont.metricValue)
+				.foregroundStyle(PanelPalette.primaryText(colorScheme))
+				.lineLimit(1)
+				.truncationMode(.middle)
+				.layoutPriority(1)
+		}
+		.frame(maxWidth: .infinity, alignment: .leading)
+	}
+}
+
+struct PanelIconButtonView: View {
+	let symbol: String
+	let tint: Color
+	let isActive: Bool
+	let isDestructive: Bool
+	let isDisabled: Bool
+	let isSubtle: Bool
+	let isPrimary: Bool
+	let size: CGFloat
+	let action: () -> Void
+	let help: String
+	@Environment(\.colorScheme) private var colorScheme
+
+	init(
+		symbol: String,
+		tint: Color,
+		isActive: Bool,
+		isDestructive: Bool = false,
+		isDisabled: Bool = false,
+		isSubtle: Bool = false,
+		isPrimary: Bool = false,
+		size: CGFloat = 24,
+		action: @escaping () -> Void,
+		help: String
+	) {
+		self.symbol = symbol
+		self.tint = tint
+		self.isActive = isActive
+		self.isDestructive = isDestructive
+		self.isDisabled = isDisabled
+		self.isSubtle = isSubtle
+		self.isPrimary = isPrimary
+		self.size = size
+		self.action = action
+		self.help = help
+	}
+
+	var body: some View {
+		Button(action: action) {
+			buttonLabel
+		}
+		.buttonStyle(
+			PanelInteractiveButtonStyle(
+				isDisabled: isDisabled,
+				hoverLift: 0,
+				hoverScale: isSubtle ? 1.004 : 1.006,
+				pressedScale: 0.952,
+				hoverShadowRadius: isSubtle ? 2.4 : 3
+			)
+		)
+		.disabled(isDisabled)
+		.opacity(isDisabled && isActive == false ? 0.56 : 1)
+		.help(help)
+	}
+
+	@ViewBuilder
+	private var buttonLabel: some View {
+		if usesSurface {
+			iconContent
+				.modernGlassSurface(
+					cornerRadius: iconCornerRadius,
+					depth: .control
+				)
+		} else {
+			iconContent
+				.opacity(isDisabled ? 0.34 : 0.82)
+		}
+	}
+
+	private var iconContent: some View {
+		Image(systemName: symbol)
+			.font(PanelFont.iconButton)
+			.symbolRenderingMode(.monochrome)
+			.foregroundStyle(foregroundColor)
+			.frame(width: size, height: size)
+			.contentShape(RoundedRectangle(cornerRadius: iconCornerRadius, style: .continuous))
+	}
+
+	private var foregroundColor: Color {
+		if isActive {
+			return tint.opacity(colorScheme == .dark ? 0.98 : 0.92)
+		}
+		if isDisabled {
+			return PanelPalette.secondaryText(colorScheme).opacity(0.38)
+		}
+		if isDestructive {
+			return tint.opacity(colorScheme == .dark ? 0.96 : 0.9)
+		}
+		if isPrimary {
+			return tint.opacity(colorScheme == .dark ? 1 : 0.96)
+		}
+		if isSubtle {
+			return tint.opacity(colorScheme == .dark ? 0.86 : 0.82)
+		}
+		return PanelPalette.actionBlue(colorScheme).opacity(colorScheme == .dark ? 0.88 : 0.86)
+	}
+
+	private var usesSurface: Bool {
+		if isSubtle {
+			return false
+		}
+		if isActive || isPrimary {
+			return true
+		}
+		return true
+	}
+
+	private var iconCornerRadius: CGFloat {
+		size * 0.5
+	}
+}
+
